chore(scripts): Defensive symlink

+++ scripts/symlink.sh
@@ -1,3 +1,11 @@
-ln -s "./dist/client-node.js" client-node.js
-ln -s "./dist/client-web.js" client-web.js
-ln -s "./dist/server.js" server.js
+if [ ! -f client-node.js ]; then
+    ln -s "./dist/client-node.js" client-node.js
+fi
+
+if [ ! -f client-web.js ]; then
+    ln -s "./dist/client-web.js" client-web.js
+fi
+
+if [ ! -f server.js ]; then
+    ln -s "./dist/server.js" server.js
+fi
